+++ index.js
@@ -33,6 +33,28 @@ module.exports = {
           ],
         },
         {
+          /* Flex Container(Parent) */
+          properties: [
+            'flex-flow',
+            'flex-direction',
+            'flex-wrap',
+            'justify-content',
+            'align-items',
+            'align-content',
+          ],
+        },
+        {
+          /* Flex Item(Children) */
+          properties: [
+            'flex',
+            'flex-grow',
+            'flex-shrink',
+            'flex-basis',
+            'order',
+            'align-self',
+          ],
+        },
+        {
           /* Color */
           properties: [
             'background',