+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ==================
 - [breaking] `declaration-block-properties-order` added `bottom`, `left` in Positioning section.
 - [breaking] `declaration-block-properties-order` added `max-width`, `min-width`, `max-height`, `min-height` in Display & Box Model section.
-
+- [breaking] `declaration-block-properties-order` added Flex Container(Parent) section.
+- [breaking] `declaration-block-properties-order` added Flex Item(Children) section.
 
 0.1.0 / 2016-09-21
 ==================